requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
distro==1.9.0
eval_type_backport==0.2.2
-e git+ssh://git@github.com/cjavdev/python-sdk-evaluation.git@a2be2b8faa4e4ab38fbee71fcfcf786b0a5235fb#egg=fern_samsara&subdirectory=fern_samsara
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mypy-extensions==1.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
-e git+ssh://git@github.com/cjavdev/python-sdk-evaluation.git@a2be2b8faa4e4ab38fbee71fcfcf786b0a5235fb#egg=samsara&subdirectory=speakeasy_samsara
six==1.17.0
sniffio==1.3.1
-e git+ssh://git@github.com/cjavdev/python-sdk-evaluation.git@a2be2b8faa4e4ab38fbee71fcfcf786b0a5235fb#egg=speakeasy_samsara&subdirectory=speakeasy_samsara
-e git+ssh://git@github.com/cjavdev/python-sdk-evaluation.git@a2be2b8faa4e4ab38fbee71fcfcf786b0a5235fb#egg=stainless_samsara&subdirectory=stainless_samsara
typing-inspect==0.9.0
typing_extensions==4.12.2
